+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Ce projet adhère au principe du [Semantic Versioning](https://semver.org/spec/v
 - Améliore l'affichage de l'onglet "Facturation" des événements pour les utilisateurs du groupe "visiteur" (#104).
 - Place le champ "sous-catégorie" en dessous du champ "catégorie" dans l'édition de matériel (#105).
 - Pré-rempli le champ "quantité" du formulaire de nouveau matériel à 1 (#106).
+- Dans le listing du matériel, ajoute un lien vers la gestion des caractéristiques spéciales.
 
 ## 0.11.0 (2021-01-14)